delete ModalProduct,done; refactor headeradapt, WIP
